+++ sandbox-web/src/main/java/com/exadel/sandbox/team5/error/ApiError.java
@@ -1,10 +1,9 @@
-package error;
+package com.exadel.sandbox.team5.error;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.http.HttpStatus;
 
 import java.util.List;
