--- sandbox-web/src/main/java/error/CustomErrorAttributes.java
@@ -1,20 +1,0 @@
-package error;
-
-import org.springframework.boot.web.error.ErrorAttributeOptions;
-import org.springframework.boot.web.servlet.error.DefaultErrorAttributes;
-import org.springframework.stereotype.Component;
-import org.springframework.web.context.request.WebRequest;
-
-import java.util.Map;
-
-@Component
-public class CustomErrorAttributes extends DefaultErrorAttributes {
-
-    @Override
-    public Map<String, Object> getErrorAttributes(WebRequest webRequest, ErrorAttributeOptions options) {
-        Map<String, Object> errorAttributes = super.getErrorAttributes(webRequest, options);
-        errorAttributes.put("newAttribute", "value");
-        return errorAttributes;
-    }
-
-}
