--- sandbox-web/src/main/java/error/MyEntityNotFoundException.java
@@ -1,12 +1,0 @@
-package error;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class MyEntityNotFoundException extends RuntimeException {
-
-    public MyEntityNotFoundException(Long id) {
-        super("Entity is not found, id="+id);
-    }
-}
